requirements.txt
Flask==0.10.1
Flask-Admin==1.0.9
Flask-Login==0.2.11
Flask-Misaka==0.3.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
gunicorn==19.1.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
misaka==1.0.2
python-slugify==0.1.0
SQLAlchemy==0.9.8
Unidecode==0.4.17
Werkzeug==0.9.6
WTForms==2.0.2
